feat(gpu-rendering): implement hashmap lookup function in glsl script

+++ VoxelEngine/gpu_hashmap_allocator.h
@@ -55,14 +55,14 @@ private:
     static constexpr K EMPTY_KEY = std::numeric_limits<K>::max();;
     static constexpr K TOMBSTONE_KEY = std::numeric_limits<K>::min();;
 
-    static inline uint64_t Hash64(uint64_t x)
+    static inline uint32_t Hash64(uint64_t x)
     {
         x ^= x >> 33;
         x *= 0xff51afd7ed558ccdULL;
         x ^= x >> 33;
         x *= 0xc4ceb9fe1a85ec53ULL;
         x ^= x >> 33;
-        return x;
+        return (uint32_t)x;
     }
 
     size_t _Hash(const K& key) const
